refactor: move user mvc files to user package
