+++ src/main/java/com/feeling/app/user/entity/User.java
@@ -1,4 +1,4 @@
-package com.feeling.app.entity;
+package com.feeling.app.user.entity;
 
 import jakarta.persistence.*;
