+++ src/main/java/com/feeling/app/user/service/UserService.java
@@ -1,7 +1,7 @@
-package com.feeling.app.service;
+package com.feeling.app.user.service;
 
-import com.feeling.app.entity.User;
-import com.feeling.app.repository.UserRepository;
+import com.feeling.app.user.entity.User;
+import com.feeling.app.user.repository.UserRepository;
 import com.feeling.app.util.JwtProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
